MCR-3677 fix MCRClassificationEditorResourceTest on Windows
