+++ mycore-base/src/test/java/org/mycore/datamodel/ifs2/MCRFileStoreTest.java
@@ -282,7 +282,6 @@ public class MCRFileStoreTest extends MCRIFS2TestCase {
     }
 
     private void sortChildren(Element parent) {
-        @SuppressWarnings("unchecked")
         List<Element> children = parent.getChildren();
         if (children == null || children.size() == 0) {
             return;
