+++ mycore-base/src/test/java/org/mycore/datamodel/ifs2/test/MCRStoreManagerTest.java
@@ -20,66 +20,17 @@ package org.mycore.datamodel.ifs2.test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import java.io.IOException;
-import java.net.URI;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystemNotFoundException;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.util.Map;
-
 import org.junit.jupiter.api.Test;
 import org.mycore.datamodel.ifs2.MCRFileStore;
-import org.mycore.datamodel.ifs2.MCRStore.MCRStoreConfig;
 import org.mycore.datamodel.ifs2.MCRStoreManager;
-
-import com.google.common.jimfs.Configuration;
-import com.google.common.jimfs.Jimfs;
 
 public class MCRStoreManagerTest {
     @Test
     public void createMCRFileStore() throws Exception {
-        StoreConfig config = new StoreConfig();
+        MCRJimfsTestStoreConfig config = new MCRJimfsTestStoreConfig(MCRStoreManagerTest.class.getSimpleName());
         MCRFileStore fileStore = MCRStoreManager.createStore(config, MCRFileStore.class);
 
         assertNotNull(fileStore, "MCRStoreManager could not create Filestore.");
     }
 
-    class StoreConfig implements MCRStoreConfig {
-
-        private final Path baseDir;
-
-        StoreConfig() throws IOException {
-            String fsName = MCRStoreManagerTest.class.getSimpleName();
-            URI jimfsURI = URI.create("jimfs://" + fsName);
-            FileSystem fileSystem = Jimfs.newFileSystem(fsName, Configuration.unix());
-            try {
-                fileSystem = FileSystems.getFileSystem(jimfsURI);
-            } catch (FileSystemNotFoundException e) {
-                fileSystem = FileSystems.newFileSystem(jimfsURI, Map.of("fileSystem", fileSystem));
-            }
-            baseDir = fileSystem.getPath("/");
-        }
-
-        @Override
-        public String getID() {
-            return "Test";
-        }
-
-        @Override
-        public String getPrefix() {
-            return getID() + "_";
-        }
-
-        @Override
-        public String getBaseDir() {
-            return baseDir.toUri().toString();
-        }
-
-        @Override
-        public String getSlotLayout() {
-            return "4-4-2";
-        }
-
-    }
 }
