+++ mycore-classeditor/src/test/java/org/mycore/frontend/classeditor/resources/MCRClassificationEditorResourceTest.java
@@ -21,6 +21,7 @@ package org.mycore.frontend.classeditor.resources;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import java.io.IOException;
 import java.text.MessageFormat;
 import java.util.Collection;
 import java.util.List;
@@ -45,6 +46,7 @@ import org.mycore.datamodel.classifications2.MCRCategoryID;
 import org.mycore.datamodel.common.MCRXMLMetadataEventHandler;
 import org.mycore.datamodel.ifs2.MCRMetadataStore;
 import org.mycore.datamodel.ifs2.MCRStoreManager;
+import org.mycore.datamodel.ifs2.test.MCRJimfsTestStoreConfig;
 import org.mycore.frontend.classeditor.json.MCRJSONCategory;
 import org.mycore.frontend.classeditor.mocks.CategoryDAOMock;
 import org.mycore.frontend.classeditor.mocks.CategoryLinkServiceMock;
@@ -87,13 +89,19 @@ public class MCRClassificationEditorResourceTest {
             MCRSessionHookFilter.class,
             MultiPartFeature.class));
 
+        String name = MCRClassificationEditorResourceTest.class.getSimpleName();
         try {
-            MCRStoreManager.createStore("ClasseditorTempStore", MCRMetadataStore.class);
-        } catch (ReflectiveOperationException e) {
-            LOGGER.error("while creating store ClasseditorTempStore", e);
-        }
+            MCRStoreManager.computeStoreIfAbsent(name, () -> {
+                try {
+                    MCRJimfsTestStoreConfig config = new MCRJimfsTestStoreConfig(name);
+                    return MCRStoreManager.buildStore(config, MCRMetadataStore.class);
+                } catch (ReflectiveOperationException | IOException e) {
+                    LOGGER.error("while creating store ClasseditorTempStore", e);
+                    fail();
+                    return null;
+                }
+            });
 
-        try {
             /* its important to set the daomock via this method because the factory could be called
              * by a previous test. In this case a class cast exception occur because MCRCategoryDAOImpl
              * was loaded. */
